backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
httpx
bcrypt
pytz
tzdata
six
pyparsing
pillow
packaging
numpy
kiwisolver
importlib_resources
fonttools
cycler
colorama
tqdm
stanio
python-dateutil
contourpy
pandas
matplotlib
holidays
cmdstanpy
prophet
